HBX-1991: Move class 'ForeignKeysInfo' from 'org.hibernate.tool.internal.reveng' to 'org.hibernate.tool.internal.reveng.reader'
