+++ orm/src/main/java/org/hibernate/tool/internal/reveng/reader/ForeignKeysInfo.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.internal.reveng;
+package org.hibernate.tool.internal.reveng.reader;
 
 import java.util.ArrayList;
 import java.util.HashMap;
